Version 2024.11.1.dev_c: - Scheduling of boiler can be set via a ha-entity - when boiler is heated bij the heat pump for room-heating then there can only be "one" heating function in an hour, therefore is a new setting introduced in the boiler-section: "boiler heated by heatpump". This setting can be "True" or "False" - the code is brought inline with PEP 8 (Style Guide for Python Code)

+++ dao/prog/graphs.py
@@ -6,26 +6,26 @@ import numpy as np
 
 def make_graph_meteo(df, file=None, show=False):
     plt.figure(figsize=(15, 10))
-    df['gr'] = df['gr'].astype(float)
-    x_axis = np.arange(len(df['tijd_nl'].values))
-    plt.bar(x_axis-0.1,  df['gr'].values, width=0.2, label="global rad")
-    plt.bar(x_axis+0.1,  df['solar_rad'].values, width=0.2, label="netto rad")
-    plt.xticks(x_axis+0.1, df['tijd_nl'].values, rotation=45)
+    df["gr"] = df["gr"].astype(float)
+    x_axis = np.arange(len(df["tijd_nl"].values))
+    plt.bar(x_axis - 0.1, df["gr"].values, width=0.2, label="global rad")
+    plt.bar(x_axis + 0.1, df["solar_rad"].values, width=0.2, label="netto rad")
+    plt.xticks(x_axis + 0.1, df["tijd_nl"].values, rotation=45)
     if file is not None:
         plt.savefig(file)
     if show:
         plt.show()
-    plt.close('all')
+    plt.close("all")
     return
 
 
 def make_graph_entsoe(df):
     plt.figure(figsize=(15, 10))
-    df['gr'] = df['gr'].astype(float)
-    x_axis = np.arange(len(df['tijd_nl'].values))
-    plt.bar(x_axis-0.1,  df['gr'].values, width=0.2, label="global rad")
-    plt.bar(x_axis+0.1,  df['solar_rad'].values, width=0.2, label="netto rad")
-    plt.xticks(x_axis, df['tijd_nl'].values, rotation=45)
+    df["gr"] = df["gr"].astype(float)
+    x_axis = np.arange(len(df["tijd_nl"].values))
+    plt.bar(x_axis - 0.1, df["gr"].values, width=0.2, label="global rad")
+    plt.bar(x_axis + 0.1, df["solar_rad"].values, width=0.2, label="netto rad")
+    plt.xticks(x_axis, df["tijd_nl"].values, rotation=45)
     plt.show()
-    plt.close('all')
+    plt.close("all")
     return
